addressing comments, fixing tests and removing the moveservice
