+++ src/main/java/uk/gov/hmcts/reform/finrem/caseorchestration/helper/DocumentHelper.java
@@ -25,6 +25,7 @@ import uk.gov.hmcts.reform.finrem.caseorchestration.service.CommonFunction;
 import java.io.IOException;
 import java.time.LocalDate;
 import java.util.ArrayList;
+import java.util.Collection;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
@@ -372,5 +373,21 @@ public class DocumentHelper {
     public List<HearingOrderCollectionData> getFinalOrderDocuments(Map<String, Object> caseData) {
         return objectMapper.convertValue(caseData.get(FINAL_ORDER_COLLECTION),
             new TypeReference<>() {});
+    }
+
+    public Map<String, Object> moveCollection(Map<String, Object> caseData, String source, String destination) {
+        if (caseData.get(source) != null && (caseData.get(source) instanceof Collection)) {
+            if (caseData.get(destination) == null || (caseData.get(destination) instanceof Collection)) {
+                final List destinationList = new ArrayList();
+                if (caseData.get(destination) != null) {
+                    destinationList.addAll((List) caseData.get(destination));
+                }
+                destinationList.addAll((List) caseData.get(source));
+                caseData.put(destination, destinationList);
+                caseData.put(source, null);
+            }
+        }
+
+        return caseData;
     }
 }
