+++ src/test/java/uk/gov/hmcts/reform/finrem/caseorchestration/helper/DocumentHelperTest.java
@@ -1,14 +1,17 @@
 package uk.gov.hmcts.reform.finrem.caseorchestration.helper;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import org.hamcrest.Matchers;
 import org.junit.Before;
 import org.junit.Test;
 import uk.gov.hmcts.reform.bsp.common.model.document.CtscContactDetails;
 import uk.gov.hmcts.reform.ccd.client.model.CallbackRequest;
 import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
+import uk.gov.hmcts.reform.finrem.caseorchestration.TestSetUpUtils;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.CaseDocument;
 
 import java.io.InputStream;
+import java.util.Collection;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -16,6 +19,7 @@ import java.util.Optional;
 
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.CoreMatchers.nullValue;
+import static org.hamcrest.Matchers.hasSize;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertThat;
 import static uk.gov.hmcts.reform.finrem.caseorchestration.OrchestrationConstants.CTSC_CARE_OF;
@@ -29,6 +33,7 @@ import static uk.gov.hmcts.reform.finrem.caseorchestration.OrchestrationConstant
 import static uk.gov.hmcts.reform.finrem.caseorchestration.TestSetUpUtils.defaultConsentedCaseDetails;
 import static uk.gov.hmcts.reform.finrem.caseorchestration.helper.DocumentHelper.CTSC_CONTACT_DETAILS;
 import static uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.CCDConfigConstant.CONSENT_ORDER;
+import static uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.CCDConfigConstant.HEARING_ORDER_COLLECTION;
 
 public class DocumentHelperTest {
 
@@ -246,6 +251,56 @@ public class DocumentHelperTest {
             .build();
 
         assertEquals(ctscContactDetails, preparedCaseDetails.getData().get(CTSC_CONTACT_DETAILS));
+    }
+
+    @Test
+    public void shouldSuccessfullyMoveValues() throws Exception {
+        Map<String, Object> caseData = TestSetUpUtils.caseDataWithUploadHearingOrder();
+
+        documentHelper.moveCollection(caseData,HEARING_ORDER_COLLECTION, "uploadHearingOrderRO");
+
+        assertThat(((Collection<CaseDocument>)caseData.get("uploadHearingOrderRO")), hasSize(3));
+        assertThat(caseData.get(HEARING_ORDER_COLLECTION), Matchers.nullValue());
+    }
+
+    @Test
+    public void shouldSuccessfullyMoveValuesToNewCollections() throws Exception {
+        Map<String, Object> caseData = TestSetUpUtils.caseDataWithUploadHearingOrder();
+        caseData.put("uploadHearingOrderRO", null);
+        documentHelper.moveCollection(caseData,HEARING_ORDER_COLLECTION, "uploadHearingOrderRO");
+
+        assertThat(((Collection<CaseDocument>)caseData.get("uploadHearingOrderRO")), hasSize(1));
+        assertThat(caseData.get(HEARING_ORDER_COLLECTION), Matchers.nullValue());
+    }
+
+    @Test
+    public void shouldDoNothingWithNonArraySourceValueMove() throws Exception {
+        Map<String, Object> caseData = TestSetUpUtils.caseDataWithUploadHearingOrder();
+        caseData.put(HEARING_ORDER_COLLECTION, "nonarrayValue");
+        documentHelper.moveCollection(caseData, HEARING_ORDER_COLLECTION, "uploadHearingOrderRO");
+
+        assertThat(((Collection<CaseDocument>)caseData.get("uploadHearingOrderRO")), hasSize(2));
+        assertThat(caseData.get(HEARING_ORDER_COLLECTION), Matchers.is("nonarrayValue"));
+    }
+
+    @Test
+    public void shouldDoNothingWithNonArrayDestinationValueMove() throws Exception {
+        Map<String, Object> caseData = TestSetUpUtils.caseDataWithUploadHearingOrder();
+        caseData.put("uploadHearingOrderRO", "nonarrayValue");
+        documentHelper.moveCollection(caseData, HEARING_ORDER_COLLECTION, "uploadHearingOrderRO");
+
+        assertThat(caseData.get("uploadHearingOrderRO"), Matchers.is("nonarrayValue"));
+        assertThat(((Collection<CaseDocument>)caseData.get(HEARING_ORDER_COLLECTION)), hasSize(1));
+    }
+
+    @Test
+    public void shouldDoNothingWhenSourceIsEmptyMove() throws Exception {
+        Map<String, Object> caseData = TestSetUpUtils.caseDataWithUploadHearingOrder();
+        caseData.put(HEARING_ORDER_COLLECTION, null);
+        documentHelper.moveCollection(caseData, HEARING_ORDER_COLLECTION, "uploadHearingOrderRO");
+
+        assertThat(((Collection<CaseDocument>)caseData.get("uploadHearingOrderRO")), hasSize(2));
+        assertThat(caseData.get(HEARING_ORDER_COLLECTION), Matchers.nullValue());
     }
 
     private CallbackRequest prepareCallbackRequestForLatestConsentedConsentOrder(String fileName) throws Exception {
